requirements.txt
streamlit
openai
elevenlabs
replicate
supabase
pandas
python-dotenv